refactor: move table exclusion logic to shared utility function
