--- src/internal/default-excluded-tables.ts
@@ -1,4 +1,0 @@
-export const DEFAULT_EXCLUDED_TABLES = [
-  'kysely_migration',
-  'kysely_migration_lock',
-]
